@@ -4,13 +4,14 @@
   "description": "code template of express application",
   "main": "index.js",
   "scripts": {
-    "start":   "node src/index.js",
-    "dev":     "node-dev src/index.js",
-    "lint":    "eslint . --config eslintrc.js",
-    "lintfix": "eslint . --config eslintrc.js --fix",
-    "test":    "jest"
+    "build": "tsc",
+    "clean": "rm -rf dist/",
+    "dev": "node-dev dist/index.js",
+    "lint": "eslint . --config eslintrc.js --ext .js,.ts",
+    "lintfix": "eslint . --config eslintrc.js --ext .js,.ts --fix",
+    "start": "node dist/index.js",
+    "test": "jest"
   },
-  "author": "mozzzzy",
   "private": "true",
   "dependencies": {
     "express": "^4.17.1",
@@ -18,11 +19,17 @@
     "log4js": "^6.3.0"
   },
   "devDependencies": {
-    "eslint": "^8.0.0",
+    "@types/express": "^4.17.13",
+    "@types/jest": "^27.4.0",
+    "@types/supertest": "^2.0.11",
+    "@typescript-eslint/eslint-plugin": "^5.12.1",
+    "@typescript-eslint/parser": "^5.12.1",
+    "eslint": "^8.9.0",
     "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-plugin-import": "^2.22.1",
-    "jest": "^27.0.0",
+    "eslint-plugin-import": "^2.25.4",
+    "jest": "^27.5.1",
     "node-dev": "^7.0.0",
-    "supertest": "^6.0.0"
+    "supertest": "^6.0.0",
+    "ts-jest": "^27.1.3"
   }
 }
